move the anomalous namespace to deeptau base
